pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name="infscale"
dynamic = ["version"]
authors = [{ name="Infscale Maintainers", email="infscale-github-owners@cisco.com" }]
description="This package is a python library to serve ML model in the infscale system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate >= 0.24.1",
    "click>=8.1.7",
    "cloudpickle >= 3.0.0",
    "datasets >= 2.14.6",
    "fastapi >= 0.109.0",
    "grpcio >= 1.60.0",
    "protobuf >= 4.25.1",
    "pydantic >= 2.4.2",
    "pynvml >= 11.5.0",
    "torch == 2.0.1",
    "torchvision >= 0.15.2",
    "transformers >= 4.30.2",
    "uvicorn >= 0.26.0"
]

[project.scripts]
infscale = "infscale.cli:cli"

[project.optional-dependencies]
dev = [
    "black",
    "bumpver",
    "grpcio-tools",
    "isort",
    "pip-tools",
    "pytest",
    "python-lsp-server",
]

[tool.setuptools.dynamic]
version = {attr = "infscale.__version__"}
